+++ astris/__init__.py
@@ -1,11 +1,11 @@
-from .app import AstrisApp
+from .app import Astris
 from .component import Component, Element, Text
 from .content import JsonCollection, JsonCollectionEntry, register_json_collection
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 
 __all__ = [
-    "AstrisApp",
+    "Astris",
     "Component",
     "Element",
     "Text",